chore: cleanup after BSP refactoring
- Extract shared utf8ToCP437() into displayCommon.cpp
- Remove orphaned T-ETH-Elite (non-SX1262) board
- Consolidate platformio.ini with [esp32_base] extends pattern
- Normalize file naming: replace hyphens with underscores in all
  hal and display file names for consistency with directory names
