+++ src/bsp/boards/LILYGO_T_ETH_Elite_SX1262/hal_LILYGO_T_ETH_Elite_SX1262.cpp
